+++ packages/tests/Base/utils.spec.ts
@@ -1,8 +1,13 @@
-import { describe, it, expect, jest } from 'bun:test';
+import { describe, it, expect, mock } from 'bun:test';
+import { Base, withName } from '@studiometa/js-toolkit';
 import { nextTick } from '@studiometa/js-toolkit/utils';
-import { getComponentElements, addToQueue } from '../../js-toolkit/Base/utils.js';
-import { features } from '../../js-toolkit/Base/features.js';
-import { h } from '../__utils__/h.js';
+import {
+  getComponentElements,
+  addToQueue,
+  registerInstance,
+  getInstances,
+} from '#private/Base/utils.js';
+import { h } from '#test-utils';
 
 describe('The `getComponentElements` function', () => {
   it('should find components with multiple declarations', () => {
@@ -32,14 +37,31 @@ describe('The `getComponentElements` function', () => {
 });
 
 describe('The `addToQueue` function', () => {
-  it('should delay given tasks if the `asyncChildren` feature is enabled', async () => {
-    features.set('asyncChildren', true);
-    const fn = jest.fn();
+  it('should delay given tasks if the `blocking` feature is disabled', async () => {
+    const map = new Map([['blocking', false]]);
+    mock.module('#private/Base/features.js', () => ({ features: map }));
+    const fn = mock();
 
     addToQueue(fn);
     expect(fn).not.toHaveBeenCalled();
     await nextTick();
     expect(fn).toHaveBeenCalledTimes(1);
-    features.set('asyncChildren', false);
+  });
+});
+
+describe('The instance helpers', () => {
+  it('should save and get instances for a given constructor', () => {
+    const Foo = withName(Base, 'Foo');
+    const instances = getInstances(Foo);
+
+    expect(instances).toBeInstanceOf(Set);
+    expect(getInstances(Foo) === instances).toBe(true);
+    expect(instances).toHaveLength(0);
+
+    const foo = new Foo(h('div'));
+
+    expect(Array.from(instances)[0]).toBe(foo);
+    expect(instances).toHaveLength(1);
+    expect(instances.has(foo)).toBe(true);
   });
 });