+++ packages/tests/__utils__/resizeWindow.ts
@@ -1,21 +1,23 @@
-import { useFakeTimers, useRealTimers, advanceTimersByTimeAsync } from './faketimers.js';
+import {
+  useFakeTimers,
+  useRealTimers,
+  advanceTimersByTimeAsync,
+  isFakeTime,
+} from './faketimers.js';
 
-/**
- * Resize the jsdom window to the given size.
- *
- * @param  {Object}  options
- * @param  {Number}  [options.width=window.innerHeight]  The new width.
- * @param  {Number}  [options.height=window.innerHeight] The new height.
- * @return {Promise}                A promise waiting longer than the debounced event from the resize service.
- */
-export default async function resizeWindow({
+export async function resizeWindow({
   width = window.innerWidth,
   height = window.innerHeight,
 } = {}) {
-  useFakeTimers();
+  const hasFakeTimer = isFakeTime();
+  if (!hasFakeTimer) {
+    useFakeTimers();
+  }
   window.innerWidth = width;
   window.innerHeight = height;
   window.dispatchEvent(new Event('resize'));
   await advanceTimersByTimeAsync(400);
-  useRealTimers();
+  if (!hasFakeTimer) {
+    useRealTimers();
+  }
 }